Check API data, cache refrest, and count display
At one point the Twitter API returned tweet counts for
URLs different to that which was requested. This resulted
in the permanent caching of high and inaccurate retweet
counts. The code was updated as follows:

- A periodic refresh of the cache is performed.

- The URL that the Twitter API count data is associated
  with is checked to match the URL of the post.

In addition, the display of the count data was made
customisable: never; when greater than zero; always.